bot_code/requirements.txt
# Core dependencies
requests<=2.32.4
asyncio<=3.4.3
aiohttp<=3.12.13
asyncio-mqtt<=0.16.2
websockets<=15.0.1
python-dotenv<=1.1.1

# Social Media APIs
tweepy<=4.16.0
praw<=7.8.1
discord.py<=2.5.2
telethon<=1.40.0
TikTokApi<=7.1.0
python-telegram-bot

# Blockchain and Trading
solana<=0.36.7
solders<=0.26.0
anchorpy<=0.21.0
construct<=2.10.70

# AI and Data Analysis
# pandas==2.2.3
# numpy==1.26.4
pandas<=2.3.0
numpy<=2.3.1
scikit-learn<=1.7.0
tensorflow<=2.19.0
torch<=2.7.1
transformers<=4.52.4

# Web Interface
flask<=3.1.1
flask-socketio<=5.5.1
flask-cors<=6.0.1

# Database
#sqlite3

# Data Processing
beautifulsoup4<=4.13.4
selenium<=4.33.0
lxml<=5.4.0

# Logging and Monitoring
loguru<=0.7.3
prometheus-client<=0.22.1

# Configuration and Security
cryptography<=45.0.4
python-jose<=3.5.0
passlib>=1.7.4

# Testing
pytest<=8.4.1
pytest-asyncio<=1.0.0
